Added attributes for object type
