+++ src/Core/Types/Types/Attributes/Argument.cs
@@ -25,24 +25,6 @@ namespace HotChocolate.Types
         AttributeTargets.Class | AttributeTargets.Struct,
         Inherited = true,
         AllowMultiple = true)]
-    public abstract class ObjectTypeDescriptorAttribute
-        : DescriptorAttribute
-    {
-        internal sealed override void TryConfigure(IDescriptor descriptor)
-        {
-            if (descriptor is IObjectTypeDescriptor d)
-            {
-                OnConfigure(d);
-            }
-        }
-
-        public abstract void OnConfigure(IObjectTypeDescriptor descriptor);
-    }
-
-    [AttributeUsage(
-        AttributeTargets.Class | AttributeTargets.Struct,
-        Inherited = true,
-        AllowMultiple = true)]
     public abstract class InputObjectTypeDescriptorAttribute
         : DescriptorAttribute
     {
