+++ src/Core/Types/Types/Attributes/ObjectTypeDescriptorAttribute.cs
@@ -1,0 +1,23 @@
+using System;
+
+namespace HotChocolate.Types
+{
+    [AttributeUsage(
+        AttributeTargets.Class | AttributeTargets.Struct,
+        Inherited = true,
+        AllowMultiple = true)]
+    public abstract class ObjectTypeDescriptorAttribute
+        : DescriptorAttribute
+    {
+        internal sealed override void TryConfigure(IDescriptor descriptor)
+        {
+            if (descriptor is IObjectTypeDescriptor d)
+            {
+                OnConfigure(d);
+            }
+        }
+
+        public abstract void OnConfigure(IObjectTypeDescriptor descriptor);
+    }
+
+}
